+++ NEWS.md
@@ -1,8 +1,12 @@
 # jsTreeR 2.4.0 (2023-11-??)
 
-- New option `tieSelection`. It is a `TRUE`/`FALSE` option about the behavior 
-of the checkboxes: it controls whether the checkboxes must be selected when
+- New option `checkWithText`. It is a `TRUE`/`FALSE` option about the behavior 
+of the checkboxes: it controls whether the checkboxes must be (un)checked when
 clicking on the text of a node.
+
+- The Shiny input value `ID_selected_tree` (see `?jstree-shiny`) has been 
+renamed to `ID_checked_tree`. If you set `checkWithText` to `FALSE`, then 
+you *must* use this input value to get the selected nodes.
 
 
 # jsTreeR 2.3.3 (2023-10-10)